+++ idp_engine/Annotate.py
@@ -24,6 +24,7 @@ from __future__ import annotations
 from copy import copy, deepcopy
 from itertools import chain
 import string
+from typing import Dict
 
 from .Parse import (Vocabulary, Import, TypeDeclaration, Declaration,
     SymbolDeclaration, VarDeclaration, TheoryBlock, Definition, Rule,
@@ -273,7 +274,7 @@ def annotate(self, voc, q_vars):
             if type(arg) == Variable \
             and arg.name in vars and arg.name not in new_vars_dict:  # a variable, but not repeated (and not a new variable name, by chance)
                 del vars[arg.name]
-                rename_args(renamed, [arg], [nv])
+                rename_args(renamed, {arg.name: nv})
             else:
                 eq = EQUALS([nv, arg])
                 renamed.body = AND([eq, renamed.body])
@@ -291,7 +292,7 @@ def annotate(self, voc, q_vars):
             if type(arg) == Variable \
             and arg.name in vars and arg.name not in new_vars:  # a variable, but not repeated (and not a new variable name, by chance)
                 del vars[arg.name]
-                rename_args(canonical, [arg], [nv])
+                rename_args(canonical, {arg.name: nv})
             else:
                 eq = EQUALS([nv, arg])
                 canonical.body = AND([eq, canonical.body])
@@ -340,16 +341,16 @@ def annotate(self, voc, q_vars):
     return self
 Rule.annotate = annotate
 
-def rename_args(self, old, new):
+def rename_args(self: Rule, subs: Dict[str, Expression]):
     """replace old variables by new variables
         (ignoring arguments in the head before the it
     """
-    self.body = self.body.instantiate(old, new)
-    self.out = (self.out.instantiate(old, new) if self.out else
+    self.body = self.body.interpret(None, subs)
+    self.out = (self.out.interpret(None, subs) if self.out else
                 self.out)
     args = self.definiendum.sub_exprs
     for j in range(0, len(args)):
-        args[j] = args[j].instantiate(old, new)
+        args[j] = args[j].interpret(None, subs)
 
 
 # Class Structure  #######################################################