+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Contributors (alphabetical): Pierre Carbonnelle, Simon Vandevelde.
 ### IDP-Z3
 * MR !319: improve some error messages
 * issue #267: Use variable declaration for type inference in recursive definitions
-*
+* issue #284: incorrect propagation when quantifying over infinite domain
 *
 
 ### Interactive Consultant and Web IDE